+++ conjurapi/policy.go
@@ -1,25 +1,43 @@
 package conjurapi
 
 import (
-	"fmt"
 	"io"
-	"net/http"
 
-	"github.com/cyberark/conjur-api-go/conjurapi/wrapper"
+	"github.com/cyberark/conjur-api-go/conjurapi/response"
 )
 
-func (c *Client) LoadPolicy(policyId string, policy io.Reader) (map[string]interface{}, error) {
-	var (
-		req *http.Request
-		err error
-	)
+// PolicyMode defines the server-sized behavior when loading a policy.
+type PolicyMode uint
 
-	if c.config.V4 {
-		err = fmt.Errorf("LoadPolicy is not supported for Conjur V4")
-	} else {
-		req, err = wrapper.LoadPolicyRequest(c.config.ApplianceURL, makeFullId(c.config.Account, "policy", policyId), policy)
-	}
+const (
+	// PolicyModePost appends new data to the policy.
+	PolicyModePost PolicyMode = 1
+	// PolicyModePut completely replaces the policy, implicitly deleting data which is not present in the new policy.
+	PolicyModePut PolicyMode = 2
+	// PolicyModePatch adds policy data and explicitly deletes policy data.
+	PolicyModePatch PolicyMode = 3
+)
 
+// CreatedRole contains the full role ID and API key of a role which was created
+// by the server when loading a policy.
+type CreatedRole struct {
+	ID     string
+	APIKey string `json:"api_key"`
+}
+
+// PolicyResponse contains information about the policy update.
+type PolicyResponse struct {
+	// Newly created roles.
+	CreatedRoles map[string]CreatedRole `json:"created_roles"`
+	// The version number of the policy.
+	Version uint32
+}
+
+// LoadPolicy submits new policy data or polciy changes to the server.
+//
+// The required permission depends on the mode.
+func (c *Client) LoadPolicy(mode PolicyMode, policyID string, policy io.Reader) (*PolicyResponse, error) {
+	req, err := c.router.LoadPolicyRequest(mode, policyID, policy)
 	if err != nil {
 		return nil, err
 	}
@@ -29,5 +47,6 @@ func (c *Client) LoadPolicy(policyId string, policy io.Reader) (map[string]inter
 		return nil, err
 	}
 
-	return wrapper.LoadPolicyResponse(resp)
+	policyResponse := PolicyResponse{}
+	return &policyResponse, response.JSONResponse(resp, &policyResponse)
 }