+++ conjurapi/policy_test.go
@@ -2,12 +2,15 @@ package conjurapi
 
 import (
 	"fmt"
-	"github.com/cyberark/conjur-api-go/conjurapi/authn"
-	"github.com/cyberark/conjur-api-go/conjurapi/response"
-	. "github.com/smartystreets/goconvey/convey"
+	"math/rand"
 	"os"
 	"strings"
 	"testing"
+	"time"
+
+	"github.com/cyberark/conjur-api-go/conjurapi/authn"
+	"github.com/cyberark/conjur-api-go/conjurapi/response"
+	. "github.com/smartystreets/goconvey/convey"
 )
 
 func TestClient_LoadPolicy(t *testing.T) {
@@ -15,35 +18,63 @@ func TestClient_LoadPolicy(t *testing.T) {
 		config := &Config{}
 		config.mergeEnv()
 
-		api_key := os.Getenv("CONJUR_AUTHN_API_KEY")
+		apiKey := os.Getenv("CONJUR_AUTHN_API_KEY")
 		login := os.Getenv("CONJUR_AUTHN_LOGIN")
 
+		conjur, err := NewClientFromKey(*config, authn.LoginPair{login, apiKey})
+		So(err, ShouldBeNil)
+
+		randomizer := rand.New(rand.NewSource(time.Now().UnixNano()))
+
 		Convey("Successfully load policy", func() {
-			variable_identifier := "alice"
+			username := "alice"
 			policy := fmt.Sprintf(`
 - !user %s
-`, variable_identifier)
-
-			conjur, err := NewClientFromKey(*config, authn.LoginPair{login, api_key})
-			So(err, ShouldBeNil)
+`, username)
 
 			resp, err := conjur.LoadPolicy(
+				PolicyModePut,
 				"root",
 				strings.NewReader(policy),
 			)
 
 			So(err, ShouldBeNil)
-			So(resp["created_roles"], ShouldNotBeNil)
+			So(resp.Version, ShouldBeGreaterThanOrEqualTo, 1)
+		})
+
+		Convey("A new role is reported in the policy load response", func() {
+			const chars = "abcdefghijklmnopqrstuvwxyz0123456789"
+			result := make([]byte, 12)
+			for i := range result {
+				result[i] = chars[randomizer.Intn(len(chars))]
+			}
+
+			username := string(result)
+			policy := fmt.Sprintf(`
+- !user %s
+`, username)
+
+			resp, err := conjur.LoadPolicy(
+				PolicyModePut,
+				"root",
+				strings.NewReader(policy),
+			)
+
+			So(err, ShouldBeNil)
+			createdRole, ok := resp.CreatedRoles["cucumber:user:"+username]
+			So(createdRole.ID, ShouldNotBeBlank)
+			So(createdRole.APIKey, ShouldNotBeBlank)
+			So(ok, ShouldBeTrue)
 		})
 
 		Convey("Given invalid login credentials", func() {
 			login = "invalid-user"
 
 			Convey("Returns 401", func() {
-				conjur, err := NewClientFromKey(*config, authn.LoginPair{login, api_key})
+				conjur, err := NewClientFromKey(*config, authn.LoginPair{login, apiKey})
 				So(err, ShouldBeNil)
 
-				resp, err := conjur.LoadPolicy("root", strings.NewReader(""))
+				resp, err := conjur.LoadPolicy(PolicyModePut, "root", strings.NewReader(""))
 
 				So(err, ShouldNotBeNil)
 				So(resp, ShouldBeNil)
@@ -63,18 +94,19 @@ func TestClient_LoadPolicy(t *testing.T) {
 		}
 
 		login := os.Getenv("CONJUR_V4_AUTHN_LOGIN")
-		api_key := os.Getenv("CONJUR_V4_AUTHN_API_KEY")
+		apiKey := os.Getenv("CONJUR_V4_AUTHN_API_KEY")
+
+		conjur, err := NewClientFromKey(*config, authn.LoginPair{login, apiKey})
+		So(err, ShouldBeNil)
 
 		Convey("Policy loading is not supported", func() {
-			variable_identifier := "alice"
+			variableIdentifier := "alice"
 			policy := fmt.Sprintf(`
 - !user %s
-`, variable_identifier)
-
-			conjur, err := NewClientFromKey(*config, authn.LoginPair{login, api_key})
-			So(err, ShouldBeNil)
+`, variableIdentifier)
 
 			_, err = conjur.LoadPolicy(
+				PolicyModePut,
 				"root",
 				strings.NewReader(policy),
 			)